Split github/project_test.go into focused test files
Move TestFetchProjectBoard_* to fetch_board_test.go, TestFetchItemDetails_*
to fetch_details_test.go. project_test.go retains TestParseTime and the
readVars/makeItem helpers used by both split files.
